Break step-sequence cycles by isolating useStepSequence
