+++ frontend/src/pages/explorateurIA/worlds/world1/steps.ts
@@ -1,7 +1,7 @@
 import {
   isCompositeStepDefinition,
   type StepDefinition,
-} from "../../../../modules/step-sequence";
+} from "../../../../modules/step-sequence/types";
 
 import {
   DEFAULT_CLARTE_QUIZ_CONFIG,
